index.htm:  Remove canvas (not needed)
Using canvas for this project was unnecessary.  Canvas is only really needed for drawing complex stuff.... in this case, all we're doing is filling and clearing little boxes, which can be done with divs alone.  I only realized the problem when I started trying to add divs to my canvas, and they weren't showing up (they're not supposed to).